Update to core log in functions
this update fixes the ability to log out from the webpage, all using AJAX and JQuery.

included in this update is lif - log in fragment - this is to be returned to the web-client upon logging out of the system.

adjustmetns have been made to some of the javascript ajax code to further the request to be logged out of the system and request the fragment of the webpage